problems/0073-set-matrix-zeroes/set_zeroes.go
package _073_set_matrix_zeroes

func setZeroes(matrix [][]int) {
	m, n := len(matrix), len(matrix[0])
	firstRow := false
	for i := 0; i < m; i++ {
		for j := 0; j < n; j++ {
			if matrix[i][j] == 0 {
				matrix[0][j] = 0
				if i == 0 {
					firstRow = true
				} else {
					matrix[i][0] = 0
				}
			}
		}
	}

	for i := 1; i < m; i++ {
		if matrix[i][0] == 0 {
			for j := 1; j < n; j++ {
				matrix[i][j] = 0
			}
		}
	}
	for j := 0; j < n; j++ {
		if matrix[0][j] == 0 {
			for i := 1; i < m; i++ {
				matrix[i][j] = 0
			}
		}
	}
	if firstRow {
		for j := 0; j < n; j++ {
			matrix[0][j] = 0
		}
	}
}
